refactor: move functions to the correct files
