+++ lib/screens/warehouse/warehouse_ws.dart
@@ -2,27 +2,25 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:messless/ws/backend_client.dart';
+import 'package:messless/ws/helper.dart';
+import 'package:messless/ws/schema/warehouse/warehouse.dart';
 
 import '../../ws/schema/company/company.dart';
+import '../../ws/schema/user/user.dart';
 
 class WarehouseWs {
-  static bool get isManagerOrHigher => roleAsInt() >= 3;
 
-  static bool get isAdmin => roleAsInt() == 5;
-
-  static int get currentCompanyId => _currentCompanyId();
-  static int? _activeCompanyId;
-
-  static Future<List<Map<String, dynamic>>> findAll() async {
-    final res = await BackendClient.service('warehouses').find();
-    _ensureStatus(res.status, {200});
-    return _asList(res.body);
+  static Future<List<Warehouse>> findAll() async {
+    final response = await BackendClient.service('warehouses').find();
+    HelperWs.ensureStatus(response.status, {200});
+    List<dynamic> jsonList = jsonDecode(response.body.toString());
+    return jsonList.map((json) => Warehouse.fromJson(json)).toList();
   }
 
-  static Future<Map<String, dynamic>> getById(int id) async {
+  static Future<Warehouse> getById(int id) async {
     final res = await BackendClient.service('warehouses').get(id);
-    _ensureStatus(res.status, {200});
-    return _asMap(res.body);
+    HelperWs.ensureStatus(res.status, {200});
+    return Warehouse.fromJson(jsonDecode(res.body.toString()));
   }
 
   static Future<void> create({
@@ -31,16 +29,16 @@ class WarehouseWs {
     required double longitude,
     int? companyId,
   }) async {
-    final resolvedCompanyId = companyId ?? activeCompanyId;
+    final resolvedCompanyId = companyId ?? HelperWs.activeCompanyId;
     final res = await BackendClient.service('warehouses').create(
       jsonEncode({
         "label": name,
         "latitude": latitude,
         "longitude": longitude,
-        "companyId": resolvedCompanyId
+        "companyId": resolvedCompanyId,
       }),
     );
-    _ensureStatus(res.status, {200, 201});
+    HelperWs.ensureStatus(res.status, {200, 201});
   }
 
   static Future<void> update({
@@ -60,167 +58,11 @@ class WarehouseWs {
       }),
     );
 
-    _ensureStatus(res.status, {200});
+    HelperWs.ensureStatus(res.status, {200});
   }
 
   static Future<void> delete(int id) async {
     final res = await BackendClient.service('warehouses').delete(id);
-    _ensureStatus(res.status, {200, 204});
-  }
-
-  static Future<List<Company>> findCompanies() async {
-    final res = await BackendClient.service('companies').find();
-
-    _ensureStatus(res.status, {200});
-
-    final decoded = jsonDecode(res.body ?? '[]');
-
-    if (decoded is! List) {
-      throw const FormatException('Expected company list');
-    }
-
-    return decoded.map((e) => Company.fromJson(e)).toList();
-  }
-
-  static Future<List<Map<String, dynamic>>> findUsers(int companyId) async {
-    final res = await BackendClient.service('users').get(companyId);
-
-    if (res.status != 200) {
-      throw HttpException('Fehler beim Laden der User. Status: ${res.status}');
-    }
-    final decoded = jsonDecode(res.body ?? '[]');
-    if (decoded is! List) {
-      throw const FormatException('Expected user list from backend');
-    }
-    return decoded.map((e) => Map<String, dynamic>.from(e as Map)).toList();
-  }
-
-  static Future<void> createCompany({
-    required String name,
-    required double latitude,
-    required double longitude,
-  }) async {
-    final res = await BackendClient.service('companies').create(
-      jsonEncode({
-        "label": name,
-        "latitude": latitude,
-        "longitude": longitude,
-      }),
-    );
-    _ensureStatus(res.status, {200, 201});
-  }
-
-  static String titleOf(Map<String, dynamic> json) {
-    final value = json['label'];
-
-    if (value is String && value.trim().isNotEmpty) {
-      return value.trim();
-    }
-
-    return 'Warehouse #${idOf(json)}';
-  }
-
-  static void setActiveCompanyId(int id) {
-    _activeCompanyId = id;
-  }
-
-  static void clearActiveCompanyId() {
-    _activeCompanyId = null;
-  }
-
-  static int get activeCompanyId {
-    if (isAdmin) {
-      if (_activeCompanyId == null) {
-        throw StateError('Admin hat keine Company ausgewählt');
-      }
-      return _activeCompanyId!;
-    }
-    return currentCompanyId;
-  }
-
-  static int idOf(Map<String, dynamic> json) {
-    final value = json['id'];
-    if (value is int) return value;
-    if (value is num) return value.toInt();
-    throw const FormatException('Warehouse id missing or invalid.');
-  }
-
-  static int? companyIdOf(Map<String, dynamic> json) {
-    final company = json['company'];
-
-    if (company is int) return company;
-
-    return null;
-  }
-
-  static int _currentCompanyId() {
-    final user =
-        BackendClient.authState.authenticatedConnection?.user as dynamic;
-
-    if (user == null) {
-      throw StateError('User not authenticated');
-    }
-
-    final company = user.company;
-
-    if (company is int) return company;
-
-    throw StateError('Invalid company format on user');
-  }
-
-  static int roleAsInt() {
-    final auth = BackendClient.authState.authenticatedConnection;
-    final user = auth?.user as dynamic;
-
-    if (user == null) return 0;
-
-    try {
-      final role = user.role;
-
-      if (role is int) return role;
-
-      if (role is String) {
-        switch (role) {
-          case 'Admin':
-            return 5;
-          case 'CompanyAdmin':
-            return 4;
-          case 'Manager':
-            return 3;
-          case 'Worker':
-            return 2;
-          case 'StageHand':
-            return 1;
-          default:
-            return 1;
-        }
-      }
-    } catch (_) {}
-
-    return 0;
-  }
-
-  static void _ensureStatus(int? status, Set<int> ok) {
-    if (status == null || !ok.contains(status)) {
-      throw StateError('Unexpected warehouse response status: $status');
-    }
-  }
-
-  static List<Map<String, dynamic>> _asList(String? body) {
-    final decoded = jsonDecode(body ?? '[]');
-    if (decoded is! List) {
-      throw const FormatException('Expected a JSON list.');
-    }
-
-    return decoded.map((e) => Map<String, dynamic>.from(e as Map)).toList();
-  }
-
-  static Map<String, dynamic> _asMap(String? body) {
-    final decoded = jsonDecode(body ?? '{}');
-    if (decoded is! Map) {
-      throw const FormatException('Expected a JSON object.');
-    }
-
-    return Map<String, dynamic>.from(decoded);
+    HelperWs.ensureStatus(res.status, {200, 204});
   }
 }
